move composition to compartment directory
