+++ vivarium/compartment/composition.py
