+++ vivarium/composites/gene_expression.py
@@ -5,7 +5,7 @@ import os
 import matplotlib.pyplot as plt
 
 from vivarium.actor.process import initialize_state
-from vivarium.actor.composition import get_derivers, get_schema
+from vivarium.compartment.composition import get_derivers, get_schema
 
 # processes
 from vivarium.processes.transcription import Transcription, UNBOUND_RNAP_KEY
@@ -154,7 +154,7 @@ def plot_gene_expression_output(timeseries, config, out_dir='out'):
 
 if __name__ == '__main__':
     from vivarium.actor.process import load_compartment, simulate_compartment
-    from vivarium.actor.composition import convert_to_timeseries
+    from vivarium.compartment.composition import convert_to_timeseries
 
     out_dir = os.path.join('out', 'tests', 'gene_expression_composite')
     if not os.path.exists(out_dir):
